==> tutorials/R_dataprocess/3_InitialGRN.Rmd <==
---
title: "3.Construction of an initial GRN"
author: "Yukai You"
date: '2024-10-10'
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### Step 1: Perform TF binding motif enrichment analysis Using RcisTarget

Input gene clusters and load data

```{r,echo=FALSE, results="hide", message=FALSE, warning=FALSE}
source("Function.R")
library(dplyr)
library(NetAct)
library(RcisTarget)
tfs <- readRDS("tfs.rds")
cluster.results= readRDS("cluster_shortname.rds")
data(motifAnnotations_hgnc)

```

We used **RcisTarget** to identify enriched TF binding motifs for each gene cluster.

```{r,eval=FALSE}
motifRankings <- importRankings("hg19-tss-centered-10kb-7species.mc9nr.feather")

motiftable <- list()
motifwgene <- list()

for(i in 1:length(cluster.results)) {
  tryCatch({
    motifEnrichmentTable_wGenes <- cisTarget(cluster.results[[i]], motifRankings,
                                             motifAnnot=motifAnnotations_hgnc)
    
    motifs_AUC <- calcAUC(cluster.results[[i]], motifRankings)
    
    # 2. Select significant motifs, add TF annotation & format as table
    motifEnrichmentTable <- addMotifAnnotation(motifs_AUC, 
                                               motifAnnot=motifAnnotations_hgnc, nesThreshold = 1)
    
    # 3. Identify significant genes for each motif
    motifEnrichmentTable_wGenes <- addSignificantGenes(motifEnrichmentTable, 
                                                       geneSets=cluster.results[[i]],
                                                       rankings=motifRankings, 
                                                       nCores=1,
                                                       method="aprox")
    
    motiftable[[i]] <- motifEnrichmentTable
    motifwgene[[i]] <- motifEnrichmentTable_wGenes
  }, error = function(e) {
    message(paste("Error in iteration", i, ":", e$message))
  })
}

motifwgenetotal=do.call(rbind,motifwgene)
motiftabletotal=do.call(rbind,motiftable)
saveRDS(motifwgenetotal, file = "motifwgenetotal.rds")
saveRDS(motiftabletotal, file = "motiftabletotal.rds")

```

#### Step 2: Build an initial GRN

We built an initial GRN of the core TFs (see Tutorial Part 2) according to the outcomes of motif enrichment analysis.

If needed, the Normalized Enrichment Score (NES) from **RCisTarget** could be used to select the most significant motifs (*e.g.*, NES \> 1).

```{r}
motifwgenetotal <- readRDS("motifwgenetotal.rds")
motiftabletotal <- readRDS("motiftabletotal.rds")
tfs=tfs # input your tf lists
lim=1
ae.wgene <- motifwgenetotal[NES > lim, ]
ae=motiftabletotal[motiftabletotal$NES>lim,]
TFhighConf=split(ae.wgene$TF_highConf,
                 ae$geneSet)
listgene=TFhighConf$geneSet
TFhighConf2=split(ae.wgene$enrichedGenes,
                  ae$geneSet)
listgene2=TFhighConf2$geneSet
genestest=genenames(listgene = listgene)
genestest2=targetnames(listgene2 = listgene2)
numbertest=getnrow(genestest,tfs=tfs)
genestest=updatalist(gene = genestest,number = numbertest,tfs = tfs)
genestest2=updatalist(gene = genestest2,number = numbertest,tfs = tfs)
tflinks.results=tflinks(numbertest,genes = genestest,genes2=genestest2)
tflinks.results=as.data.frame(tflinks.results)
dtotal.unique=distinct(tflinks.results, .keep_all = TRUE)
dtotal.unique[,3]=1
plot_network(dtotal.unique)
```

#### Step 3: Expand the GRN using TRRUST and/or NetAct TF-target gene databases

Additional TF-TF interactions are incorporated into the initial GRN using a TF-target gene databases. Here shows an example to use TRRUST.

```{r}

trrust_ <- read.table('trrust_rawdata.human.tsv', sep='\t', header=F)
colnames(trrust_)=c('Source','Target','Interaction','n')
trrust_tf=trrust_[trrust_[,1]%in%tfs,]
trrust_tf_2=trrust_tf[trrust_tf[,2]%in%tfs,]  
trrust_tf_3=trrust_tf_2[!(trrust_tf_2[,1]==trrust_tf_2[,2]),]
interactions=unique(trrust_tf_3[,1:2])
inter=cbind(interactions,1)
colnames(inter)=colnames(dtotal.unique)

dtotal.unique=rbind(dtotal.unique,inter)
dtotal.unique=unique(dtotal.unique)

dtotal.unique <- dtotal.unique[dtotal.unique$Source != dtotal.unique$Target, ]
dtotal_final=dtotal.unique[c(-47,-52),]
plot_network(dtotal_final)
write.csv(dtotal_final, "InitialGRN.csv")

```

The output 'InitialGRN.csv' contains the topology (*i.e.*, information regarding source, target, interaction for each edge) of the initial GRN. At this stage, no activating/inhibiting nature of each edge is specified, so the "interaction" values are set to be all 1.

Another example where interaction data from NetAct are incorporated into the initial GRN.

```{r}
data(hDB)
TF_hDB=hDB[tfs]
for (i in 1:length(TF_hDB)){
  TF_hDB[[i]]=TF_hDB[[i]][TF_hDB[[i]]%in%tfs]
}

inter_NATACT=matrix(NA,nrow=1,ncol=3)
inter_NATACT[1,]=c('Source','Target','Interaction')
for (i in 1:length(TF_hDB)){
  if (length(TF_hDB[[i]])>0){
    interaction=matrix(NA,nrow=length(TF_hDB[[i]]),ncol=3)
    interaction[,1]=names(TF_hDB[i])
    interaction[,2]=TF_hDB[[i]]
    interaction[,3]=1
    inter_NATACT=rbind(inter_NATACT,interaction)
  }
}
colnames(inter_NATACT)=inter_NATACT[1,]
inter_NATACT=inter_NATACT[-1,]
inter_NATACT=as.data.frame(inter_NATACT)
dtotal_final2=unique(rbind(dtotal_final,inter_NATACT))
write.csv(dtotal_final2, "InitialGRN2.csv")
```
