==> tutorials/R_dataprocess/5_Coarse_graining.Rmd <==
---
title: "5.GRN coarse-graining"
author: "Yukai You"
date: '2024-10-11'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### Step 1: Loading the GRN for network coarse-graining

Network coarse-graining can be performed by applying **SacoGraci** package.

Start by loading the gene regulatory network (GRN) to perform coarse-graining. The user input is a GRN in the format of network topology (*i.e.*, information regarding source, target and interaction type for each edge).

```{r}
library(readr)
library(SacoGraci)
Network_results<- read_csv("~/Desktop/Singlecell6.25/Finally fitting /network/network_828_noadd_33.csv") 
for( i in 1:nrow(Network_results)){
  if (Network_results[i,3]>1){Network_results[i,3]=1}
  else if (Network_results[i,3]<1){Network_results[i,3]=2}
  else{Network_results[i,3]=0}
}

Network_results=as.data.frame(Network_results[,1:3])
colnames(Network_results)=c("Source","Target","Type")
```

#### Step 2: RACIPE simulation of the large GRN

```{r,eval=FALSE}
data_exp = gen_RACIPE(Network_results, nModels=1000) ## nModels same to the modelsCGr in opt_SA
saveRDS(data_exp,file="data_exp.rds")
```

#### Step 3: Gene clustering

After running the RACIPE simulations, one can cluster genes using the simulated data by either:

-   **Applying a clustering algorithm**, or

-   **Manually providing gene grouping**

Below shows an example where gene grouping is provided based on visual inspection of simulated gene expression patterns from RACIPE simulations of the large GRN.

```{r}
data_exp=readRDS("data_exp.rds")
#gen_heatmap_hca(logscData = data_exp) 
gene_list = list()
gene_list[[1]]=c("FOS","HES1","LEF1")
gene_list[[2]]=c("POU5F1","TWIST2","SOX2","MYCN","E2F4")
gene_list[[3]]=c("GATA3","RARA","ZEB1","KLF8")

##output_gene_grouping = geneClustInd(logscData = data_reordered,numbGeneClust = 3)
#shG = c(1,2,3)  
#output_processed = reordering(logscData = data_reordered, gene_list = output_gene_grouping, geneGroupOrder = shG) # 
#data_processed = output_processed$data
#gene_list = output_processed$gene_list
```

#### Step 4: Inference of the coarse-grained circuits

To infer coarse-grained circuits, one needs to first determine how many clusters are present, as specified in the `numbClust` argument in the `modClustHCA` function. See **SacoGraci** documentation for more details.

```{r,eval=FALSE}

output_model_grouping = modClustHCA(logscData = data_exp,numbClust = 3)  # model clustering with HCA, we specify 3 clusters
data_reordered = output_model_grouping$dataRearr
clusterRef = output_model_grouping$clInd  # cluster indices of all models

output_processed = reordering(logscData = data_reordered, gene_list = gene_list) # reordered gene expression data by the clustering outcome
data_processed = output_processed$data
stat_clusters <-centMedVarCutDistPerc(data = data_processed, clusterRef = clusterRef, percThr = 0.01)
inTopsM <-gaInitial_gen(circuit_top = Network_results, gene_list = gene_list, numbNewTop = 90)
set.seed(2) 
circuit = opt_SA(network_top = Network_results, data = data_processed, clusterRef = clusterRef, 
                  cenMedRef = stat_clusters$center, cutOffM = stat_clusters$radius, 
                  gene_list = gene_list, init_top = inTopsM[1,], 
                  output = "Results", nRepeat= 1, modelsCGr = 100, 
                  maxT=400, threshT=40, decayRate1=0.8, decayRate2=0.6, iter_per_temp=3)## modelsCGr same to the nModels in gen_RACIPE
write.csv(circuit, "circuit.csv", row.names = FALSE)
```

##### Visualizing the optimized coarse-grained circuits

```{r}
circuit=read_csv("circuit.csv")
circuit$Source <- sapply(circuit$Source, function(x) paste(gene_list[[x]], collapse = ", "))
circuit$Target <- sapply(circuit$Target, function(x) paste(gene_list[[x]], collapse = ", "))
plot_network(circuit)
```
